show correct peakFlowmetryBounds

+++ app/src/main/java/com/example/jjp/astma/data/PeakFlowmetryBounds.kt
@@ -1,3 +1,3 @@
 package com.example.jjp.astma.data
 
-class PeakFlowmetryBounds(val topBound: Double, val bottomBound: Double, val userRate: Double)
+class PeakFlowmetryBounds(val topBound: Float, val bottomBound: Float, val userRate: Float)
